document re-compile for template change
So far the re-compilation of documents only happened when the original
CommonMark document has changed. This was fine because when the template
is updated, the user can force the re-compile by deleting the output
directory. However the web server will now correctly reflect the
If-Modified-Since request header, and the server would be lying if it
says "304 Not Modified" when, in fact, the template has been modified.
The template's mtime did not affect the compiled HTML at all.

With this commit, the mtime of the compiled HTML will be set to the
later one of the original document's mtime and the template's.