split plate-compile and plate
Combining plate-compile and plate into a single executable turned out to
be a bad idea for deployment, as it increases the size of the web server
executable. We intend to keep the server load as minimal as possible,
and a larger server app is definitely against it. Besides they do not
share any code at all anyway.