have "full" and "minimal" executables
A little bit of background:

First there were two completely independent executables. Then they were
unified into a single multimode executable for convenience and coolness.
Then it was reverted because it meant we had a bigger server-side
executable for features that won't be needed by a server.

It seems at the moment that the best solution is to have a unified
executable for local development and testing, and another executable
that is a subset, a stripped-out version solely for web serving.

This way users will only need to download one executable per system: the
full-featured `plate` for local development and testing, and the minimal
`plate-server` for server deployment.