stop erroneous behaviour with reverse proxy
Emergency bug fix. RFC 2616 dictates that redirect URIs be absolute.
This is impossible without the host name. There is no way to reliably
determine the host name behind the reverse proxy. A command line option
is probably needed to conform to RFC. In the meantime, we shall disable
the auto redirect feature.